requirements.txt
python-telegram-bot==13.7
wikipedia==1.4.0